+++ src/webapp-builtins/tests/StaticFiles.test.js
@@ -1,6 +1,11 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
+import fs from 'node:fs/promises';
+import { tmpdir } from 'node:os';
+import process from 'node:process';
+
+import { WallClock } from '@this/clocky';
 import { PathKey } from '@this/collections';
 import { MockRootComponent } from '@this/compy/testing';
 import { DispatchInfo, FullResponse } from '@this/net-util';
@@ -187,6 +192,34 @@ describe('_impl_handleRequest()', () => {
     });
   });
 
+  // Extra tests when `notFoundPath` is used
+  describe('with `notFoundPath`', () => {
+    test('notices when the file at `notFoundPath` is updated', async () => {
+      const notFoundPath = `${tmpdir()}/not-found-${process.pid}.txt`;
+      await fs.writeFile(notFoundPath, 'original');
+
+      const sf = await makeInstance({ notFoundPath });
+
+      const request  = RequestUtil.makeGet('/this/is/not/found');
+      const dispatch = new DispatchInfo(PathKey.EMPTY, request.pathname);
+
+      const result1 = await sf.handleRequest(request, dispatch);
+      const body1   = result1._testing_getBody();
+      expect(result1.status).toBe(404);
+      expect(body1.type).toBe('buffer');
+      expect(body1.buffer.toString()).toMatch(/original/);
+
+      await WallClock.waitForMsec(100); // So that `mtime` will be different.
+      await fs.writeFile(notFoundPath, 'replacement');
+
+      const result2 = await sf.handleRequest(request, dispatch);
+      const body2   = result2._testing_getBody();
+      expect(result2.status).toBe(404);
+      expect(body2.type).toBe('buffer');
+      expect(body2.buffer.toString()).toMatch(/replacement/);
+    });
+  });
+
   test.each`
   uriPath
   ${'/subdir2'}
@@ -279,6 +312,18 @@ describe('_impl_start()', () => {
 
     const sf = new StaticFiles({
       notFoundPath:  '/florp/zonk',
+      siteDirectory: STATIC_SITE_DIR
+    });
+
+    await expect(root.addAll(sf)).toReject();
+  });
+
+  test('throws given a non-existent `notFoundPath` whose directory _does_ exist', async () => {
+    const root = new MockRootComponent();
+    await root.start();
+
+    const sf = new StaticFiles({
+      notFoundPath:  `${STATIC_SITE_DIR}/what-in-the-muffin-is-happening-here.txt`,
       siteDirectory: STATIC_SITE_DIR
     });